# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/portal-openapi.yaml
openapi: 3.0.0
info:
    title: Gravitee.io Portal Rest API
    description: API dedicated to the devportal part of Gravitee
    termsOfService: terms
    contact:
        name: The Gravitee Team
        url: https://gravitee.io
        email: contact@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0
    version: "4.11.0-SNAPSHOT"
servers:
    - url: /portal/environments/{envId}
      description: The portal API for a given environment
      variables:
          envId:
              default: DEFAULT
              description: ID of the environment
...
security:
    - BasicAuth: []
    - CookieAuth: []
paths:
    /apis:
        get:
            tags:
                - Api
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
                - $ref: "#/components/parameters/apiContextPathQueryParam"
                - $ref: "#/components/parameters/apiLabelQueryParam"
                - $ref: "#/components/parameters/apiVersionQueryParam"
                - $ref: "#/components/parameters/apiNameQueryParam"
                - $ref: "#/components/parameters/apiCategoryQueryParam"
                - $ref: "#/components/parameters/apiFilterQueryParam"
                - $ref: "#/components/parameters/apiFilterExcludedQueryParam"
                - $ref: "#/components/parameters/apiPromotedQueryParam"
            summary: List APIs
            description: |
                List public APIs for anonymous requests.
                List all the APIs the current user is allowed to category for authenticated requests.

                The list can be filtered according to query parameters.

                By default the list is sorted by alphabetic order.
                If a **cat** query param is sent, a specific sorting can be applied. Please refer to **cat** description for detail.
            operationId: getApis
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List accessible APIs for current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApisResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /applications:
        get:
            tags:
                - Application
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
                - $ref: "#/components/parameters/forSubscriptionParam"
                - $ref: "#/components/parameters/applicationOrderQueryParam"
            summary: List all the applications accessible to authenticated user. Default order is by *name* ASC.
            description: |
                List all the applications accessible to authenticated user.

                User must have MANAGEMENT_APPLICATION[READ] and PORTAL_APPLICATION[READ] permission.
                For performance reasons, calling `/applications?size=-1` does not return the number of subscribers and the only information available in the owner of each application is:
                  - id
                  - displayName
                  - email
            operationId: getApplications
            responses:
                200:
                    description: List of Applications.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApplicationsResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
        post:
            tags:
                - Application
            requestBody:
                description: Use to create an application.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ApplicationInput"
            summary: Create an application
            description: |
                Create an application.

                User must have MANAGEMENT_APPLICATION[CREATE] permission.
            operationId: createApplication
            responses:
                201:
                    description: Created application.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Application"
                400:
                    description: Bad Request if body is empty.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /subscriptions:
        get:
            tags:
                - Subscription
            parameters:
                - $ref: "#/components/parameters/apiIdQueryParam"
                - $ref: "#/components/parameters/applicationIdQueryParam"
                - $ref: "#/components/parameters/subscriptionStatusesQueryParam"
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
            description: |
                List all ACCEPTED, PAUSED & PENDING subscriptions, filtered by api and/or by application. At least an api or an application must be provided.

                User must have the APPLICATION_SUBSCRIPTION[READ] permission to list subscription with application query param.\
                User must have the API_SUBSCRIPTION[READ] permission to list subscription with api query param.

            operationId: getSubscriptions
            responses:
                200:
                    description: List of Subscriptions
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubscriptionsResponse"
                400:
                    description: At least an api or an application must be provided.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
        post:
            tags:
                - Subscription
            requestBody:
                description: Use to create a subscription.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SubscriptionInput"
            summary: Create a subscription.
            description: |
                Create a new subscription. If the application used to subscribe has no api key mode (UNSPECIFIED) and the plan is APIKEY, then the api key mode can be set during creation.
                Otherwise, the api_key_mode field is ignored.

                User must have APPLICATION_SUBSCRIPTION[CREATE] permission.
            operationId: createSubscription
            responses:
                200:
                    description: Created subscription.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                400:
                    description: Input must not be null.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /user:
        get:
            tags:
                - User
            summary: Get the authenticated user
            description: |
                Get information about the authenticated user.
            operationId: getCurrentUser
            responses:
                200:
                    description: The current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                500:
                    $ref: "#/components/responses/InternalServerError"
        put:
            tags:
                - User
            requestBody:
                description: Use to update a user.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserInput"
            summary: Modify current user information.
            description: |
                Modify current user information.

                Only the current user can modify his/her information.
                The user email is not modifiable.
            operationId: updateCurrentUser
            responses:
                200:
                    description: Updated user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                400:
                    description: Input must not be null.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    description: Not allowed to modify another user.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /permissions:
        get:
            tags:
                - Permissions
            parameters:
                - $ref: "#/components/parameters/apiIdQueryParam"
                - $ref: "#/components/parameters/applicationIdQueryParam"
            summary: Retrieve user's permissions
            description: |
                Retrieve user's permissions.
            operationId: getCurrentUserPermissions
            responses:
                200:
                    description: List of user's permissions
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PermissionsResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
    /tickets:
        post:
            tags:
                - Portal
            requestBody:
                description: a new ticket to create
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TicketInput"
            summary: Create a ticket.
            description: |
                Create a ticket. This ticket can aim :
                  * a specific application
                  * a specific API
                  * the gravitee portal

                User must be authenticated to use this service.
            operationId: createTicket
            responses:
                201:
                    description: Created
                500:
                    $ref: "#/components/responses/InternalServerError"
        get:
            tags:
                - Portal
            parameters:
                - $ref: "#/components/parameters/apiIdQueryParam"
                - $ref: "#/components/parameters/applicationIdQueryParam"
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
                - $ref: "#/components/parameters/orderQueryParam"
            description: |
                List all tickets written by current user

            operationId: getTickets
            responses:
                200:
                    description: List of Tickets
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TicketsResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /configuration:
        get:
            tags:
                - Portal
            summary: Get portal configuration.
            operationId: getPortalConfiguration
            description: |
                Get all the portal configuration from the platform settings.
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: Portal configuration
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConfigurationResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
    /configuration/identities:
        get:
            tags:
                - Portal
            summary: Get the identity provider list.
            operationId: getPortalIdentityProviders
            description: |
                Get all the identity providers that can used in the portal.
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List of identity providers
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConfigurationIdentitiesResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /notifiers:
        get:
            tags:
                - Notifiers
            summary: Get notifiers list
            operationId: getNotifiers
            description: |
                Get notifiers
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: Notifier information
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Notifier"
                500:
                    $ref: "#/components/responses/InternalServerError"
    /entrypoints:
        get:
            tags:
                - Entrypoints
            parameters:
                - $ref: "#/components/parameters/connectorQueryParam"
            summary: List Entrypoints
            description: |
                List available entrypoints.
            operationId: getEntrypoints
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List available entrypoints for current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConnectorsResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
components:
...
    parameters:
...
        apiIdQueryParam:
            name: apiId
            in: query
            required: false
            description: Id of an api.
            schema:
                type: string
        applicationIdQueryParam:
            name: applicationId
            in: query
            required: false
            description: Id of an application.
            schema:
                type: string
...
        pageNumberParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        pageSizeParam:
            name: size
            in: query
            required: false
            description: |
                The number of items per page for pagination.
                If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
                If the size is -1, the response contains all datas.
            schema:
                type: integer
                default: 10

        #Apis filtering
        apiContextPathQueryParam:
            name: context-path
            in: query
            required: false
            description: The context-path of an API.
            schema:
                type: string
        apiLabelQueryParam:
            name: label
            in: query
            required: false
            description: One of the labels of an API.
            schema:
                type: string
        apiVersionQueryParam:
            name: version
            in: query
            required: false
            description: The version of an API.
            schema:
                type: string
        apiNameQueryParam:
            name: name
            in: query
            required: false
            description: The name of an API.
            schema:
                type: string
        apiCategoryQueryParam:
            name: category
            in: query
            required: false
            description: One of the category linked with an API.
            schema:
                type: string
        apiFilterQueryParam:
            name: filter
            in: query
            required: false
            description: |
                Name of a FILTER. A filter is used to pre-select some APIs. It can be combined with others query params.
                  - FEATURED : top APIs. Sorted by specific order.
                  - MINE : APIs subscribed by the current user. Sorted by alphabetical order.
                  - STARRED: Rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
                  - TRENDINGS: Subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
            schema:
                $ref: "#/components/schemas/FilterApiQuery"
        apiFilterExcludedQueryParam:
            name: -filter
            in: query
            required: false
            description: |
                Name of a FILTER to exclude. Excluded filter has priority over filter. A filter is used to pre-select some APIs. It can be combined with others query params.
                  - FEATURED : not in top APIs. Sorted by specific order.
                  - MINE : APIs not subscribed by the current user. Sorted by alphabetical order.
                  - STARRED: not rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
                  - TRENDINGS: not subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
            schema:
                $ref: "#/components/schemas/FilterApiQuery"
        apiPromotedQueryParam:
            name: promoted
            in: query
            required: false
            description: |
                Allow to filter on promoted APIs or not.
                  - If absent, no effect on the result
                  - If *false*, removes the promoted API of the list
                  - If *true*, returns only the promoted API of the list

                A promoted API is either the highlighted API of a category (if it exists) or the first API in the list.
            schema:
                type: boolean
...
        forSubscriptionParam:
            name: forSubscription
            in: query
            required: false
            description: If true, only return applications with APPLICATION_SUBSCRIPTION[CREATE] permission.
            schema:
                type: boolean
        applicationOrderQueryParam:
            name: order
            in: query
            required: false
            description: |
                By default, sort is ASC. If *field* starts with '-', the order sort is DESC.\
                Currently, only **name** and **nbSubscriptions** are supported.
            example: "order=-nbSubscriptions"
            schema:
                type: string

        #Pages filtering
...
    schemas:
...
        ErrorResponse:
            properties:
                errors:
                    description: List of errors
                    type: array
                    items:
                        $ref: "#/components/schemas/Error"
...
        ApisResponse:
            properties:
                data:
                    description: List of API.
                    type: array
                    items:
                        $ref: "#/components/schemas/Api"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        ApplicationsResponse:
            properties:
                data:
                    description: List of applications.
                    type: array
                    items:
                        $ref: "#/components/schemas/Application"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        ConnectorsResponse:
            description: List of Connectors.
            properties:
                data:
                    type: array
                    items:
                        $ref: "#/components/schemas/Connector"
...
        ApplicationInput:
            properties:
                name:
                    type: string
                description:
                    type: string
                domain:
                    type: string
                picture:
                    type: string
                groups:
                    type: array
                    items:
                        type: string
                settings:
                    $ref: "#/components/schemas/ApplicationSettings"
                background:
                    type: string
                api_key_mode:
                    $ref: "#/components/schemas/ApiKeyModeEnum"
...
    responses:
        InternalServerError:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        PermissionError:
            description: Insufficient rights to access this resource
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM